refactor: Move classes into a new aggregate package.
