+++ src/main/java/com/synopsys/integration/detect/workflow/blackduck/developer/aggregate/RapidScanComponentGroupDetail.java
@@ -5,12 +5,12 @@
  *
  * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
  */
-package com.synopsys.integration.detect.workflow.blackduck.developer;
+package com.synopsys.integration.detect.workflow.blackduck.developer.aggregate;
 
 import java.util.LinkedHashSet;
 import java.util.Set;
 
-class RapidScanComponentGroupDetail {
+public class RapidScanComponentGroupDetail {
     private RapidScanDetailGroup group;
     private Set<String> errorMessages = new LinkedHashSet<>();
     private Set<String> warningMessages = new LinkedHashSet<>();
